+++ db/migrate/20260113153953_add_erased_at_to_foi_attachments.rb
@@ -1,0 +1,5 @@
+class AddErasedAtToFoiAttachments < ActiveRecord::Migration[8.0]
+  def change
+    add_column :foi_attachments, :erased_at, :datetime
+  end
+end